serving/triton/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain<=0.3.26",
    "langchain-community<=0.3.27",
    "langchain-core<=0.3.68",
    "langchain-nvidia-ai-endpoints>=0.3.10",
    "langchain-openai<=0.3.27",
    "pillow>=11.3.0",
    "tritonclient[all]>=2.59.0",
]
